Limit deleting experiments and projects
Only project owner can delete a project

Only project owner or experiment owner can delete an experiment

+++ website/src/app/project/settings/mc-project-settings.component.js
@@ -2,13 +2,15 @@ class MCProjectSettingsComponentController {
     /*@ngInject*/
 
     /*@ngInject*/
-    constructor(projectsAPI, mcprojstore, toast, $mdDialog, $state) {
+    constructor(projectsAPI, mcprojstore, toast, $mdDialog, $state, User) {
         this.projectsAPI = projectsAPI;
         this.mcprojstore = mcprojstore;
         this.toast = toast;
         this.$mdDialog = $mdDialog;
         this.$state = $state;
         this.projectName = "";
+        this.user = User.u();
+
     }
 
     $onInit() {
@@ -34,21 +36,25 @@ class MCProjectSettingsComponentController {
     }
 
     deleteProject() {
-        let deleteDialog = this.$mdDialog.confirm()
-            .title(`Delete project: ${this.project.name}`)
-            .textContent('Deleting a project is a permanent operation - all information with the project will be removed.')
-            .ariaLabel('Delete Project')
-            .ok('Delect Project')
-            .cancel('cancel');
+        if (this.user !== this.project.owner) {
+            this.toast.error('Only the owner of a project can delete the project.')
+        } else {
+            let deleteDialog = this.$mdDialog.confirm()
+                .title(`Delete project: ${this.project.name}`)
+                .textContent('Deleting a project is a permanent operation - all information with the project will be removed.')
+                .ariaLabel('Delete Project')
+                .ok('Delect Project')
+                .cancel('cancel');
 
-        this.$mdDialog.show(deleteDialog).then(
-            () => {
-                this.projectsAPI.deleteProject(this.project.id).then(
-                    () => this.mcprojstore.removeCurrentProject().then(() => this.$state.go('projects.list')),
-                    () => this.toast.error('Failed to delete project')
-                )
-            }
-        );
+            this.$mdDialog.show(deleteDialog).then(
+                () => {
+                    this.projectsAPI.deleteProject(this.project.id).then(
+                        () => this.mcprojstore.removeCurrentProject().then(() => this.$state.go('projects.list')),
+                        () => this.toast.error('Failed to delete project')
+                    )
+                }
+            );
+        }
     }
 
     cancel() {
